Re-organize pkg into generators pkg
