+++ pkg/generators/blueprints/scaffold/misc/factory.go.tpl
