+++ pkg/generators/blueprints/scaffold/misc/mutation.graphql.tpl
