+++ pkg/generators/blueprints/scaffold/misc/query.graphql.tpl
