+++ pkg/generators/blueprints/scaffold/misc/resolver.go.tpl
