+++ pkg/generators/blueprints/scaffold/pkg/create.go.tpl
