+++ pkg/generators/blueprints/scaffold/pkg/create_test.go.tpl
