+++ pkg/generators/blueprints/scaffold/pkg/delete.go.tpl
