+++ pkg/generators/blueprints/scaffold/pkg/delete_test.go.tpl
