+++ pkg/generators/blueprints/scaffold/pkg/pkgName_suite_test.go.tpl
