+++ pkg/generators/blueprints/scaffold/pkg/schema.graphql.tpl
