+++ pkg/generators/blueprints/scaffold/pkg/service.go.tpl
