+++ pkg/generators/scaffolding/parser.go
